AddDialog: Remove separate option to put PV on an MD RAID
It is now possible to select 'lvmpv' from filesystem chooser so
there is no need for a special option for PVs on raid.

Also temporarily remove/hide 'type' selection for mdraid -- only
single filesystem option is now available so it makes no sense to
let user to choose. This selection will return with implementation
of partition table on raid option.