Split blivet-gui package into "blivet-gui" and "blivet-gui-runtime"
Anaconda now depends on blivet-gui, but Fedora Workstation group
doesn't want blivet-gui to be on the installed system. And because
Anaconda can't remove packages after installing from a LiveCD, we
need to split blivet-gui package into a "-runtime" package that
contains all the code and "blivet-gui" package that contains
only the desktop file and appdata (and depends on the "-runtime"
package).

Resolves: rhbz#1449752