Do not allow to set mountpoints for devices with children (#1667644)
It shouldn't be possible to set mountpoint for a (btrfs formatted)
partition with a btrfs volume on top of it -- mountpoint must be
set for the btrfs volume itself. There is an exception for the
subvolumes -- apparently is completely ok to set mounpoint for
both the volume and its subvolumes.