Count format minimal size when checking encrypted device min size
We still need to check min size limit for the device we are
creating to be sure to include the filesystem/format minimal size.
E.g. we need to set 10 MiB for lvmpv (2 MiB for LUKS metadata +
8 MiB for PV metadata).
This was discovered thanks to recent changes in blivet -- all
filesystems now have non-zero minimal size.