Specify starting sector when creating partitions (#1755813)
Sorting using weight unfortunately works only for non-existing
partitions and breaks if there is a preexisting partition in the
middle of the disk. The best way to be sure the partition is
always created in the correct free region selected from the GUI
seems to be specifying starting sector for each partition.
We also need to make sure the starting sector is correctly aligned
because blivet doesn't align partitions with the start sector set.
The alignment code is inspired by the `bd_part_create_part`
function from libblockdev and alignment code there.