tests: Import Gtk directly before we need it
Instead of importing it at the beginning of the file which causes
the tests to segfault when not running with an X server.

This happens only on Debian 10 and it's a result of the new
run_tests script which uses 'discover' to collect our tests even
when running only the utils test suite which otherwise doesn't
require X server to run.