Fix displaying current size of devices in the resize dialog
Blivet enforces strict alignment constraints when resizing devices
and it is possible to have a device (unaligned partition usually)
that is bigger (or smaller) than max (min) size calculated by
blivet. For these devices we can't set the current size in the
resize dialog so just set the max (min) size instead, it shouldn't
be more than few KiBs different.