Refactor checking for device resizability
We currently decide if the device is resizable completely in the
GUI using the "_allow_resize_device" function which makes some of
the devices not resizable even if they can be resized. To properly
decide we usually need to call "update_size_info" on the format
first.
This change moves the decision to the backend (except for devices
that clearly can't be resized, like protected or unsupported
devices) and shows the reason for the device not being resizable
in the resize dialog.