Fix ValueError when trying to set both upper and lower size limits
When changing size limits in the add dialog from something like
(1 MiB, 2 MiB) to (256 MiB, 1 GiB) blivet-gui crashes because we
try to change the lower limit to value greater than the upper
limit which normally doesn't make sense when changing only one of
the limits but is valid when changing both.
This change adds a new function to change both limits with one
call that makes sure the limits are changed in the correct order.

Resolves: rhbz#1881472